requirements.txt
python-telegram-bot==20.6
yt-dlp==2024.8.6
imageio-ffmpeg==0.4.9
ffmpeg-python==0.2.0
librosa==0.10.2.post1
numpy==1.26.4
scipy==1.11.4
soundfile==0.12.1
audioread==3.0.1
pydub==0.25.1
requests==2.32.3
python-dotenv==1.0.1
pyacoustid==1.3.0
tqdm==4.66.4